dual_gpu_optimizer/requirements.txt
pynvml>=11.5.0
pystray>=0.20.0
pillow>=10.3.0
tomli-w>=1.0.0
pytest>=8.2.0
typer>=0.9.0
colorama>=0.4.6
rich>=13.7.0 